changed directory structure and added new cmakelists
Changed the directory structure to something a little bit more sane. It should make things easier to work with in the future.

I also fixed some of the cmake files so that we can get everythign to compile. Still need to work on the resource files.

Still need to separate out some of the different functionality from each of the various sub-projects. Anything related to the gui should be kept completely separate from the function.